backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.5
python-dateutil==2.9.0
yfinance==0.2.50
pandas==2.2.3
numpy==2.2.1
python-multipart==0.0.20
anthropic==0.42.0
python-dotenv==1.0.0
requests==2.32.3
